+++ src/asymmetric/mod.rs
@@ -1,20 +1,14 @@
 use crate::{symmetric_crypto::SymmetricCrypto, Error, KeyTrait};
-use rand_core::{CryptoRng, RngCore};
-use std::{convert::TryFrom, vec::Vec};
+use std::{vec::Vec};
 
 pub mod ristretto;
 
-pub trait KeyPair: TryFrom<Vec<u8>> + Into<Vec<u8>> {
+pub trait KeyPair {
     /// Public key
     type PublicKey: KeyTrait;
 
     /// Private key
     type PrivateKey: KeyTrait;
-
-    /// Generate a new private key / public key couple.
-    ///
-    /// - `rng` : secure random number generator
-    fn new<R: RngCore + CryptoRng>(rng: &mut R) -> Self;
 
     /// Return a reference to the public key.
     fn public_key(&self) -> &Self::PublicKey;