+++ src/hybrid_crypto/mod.rs
@@ -13,7 +13,7 @@ mod kem;
 mod scanner;
 
 pub use block::Block;
-pub use header::{Header, Metadata};
+pub use header::{Metadata};
 use rand_core::{CryptoRng, RngCore};
 pub use scanner::BytesScanner;